requirements.txt
altgraph
annotated-types
anthropic
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asttokens
async-lru
async-timeout
attrs
Babel
backoff
beautifulsoup4
bleach
branca
cachetools
certifi
cffi
charset-normalizer
cobble
colorama
comm
contourpy
coredis
cycler
debugpy
decorator
defusedxml
Deprecated
distro
Django
docstring_parser
docx
docx2txt
elastic-transport
elasticsearch
et-xmlfile
exception
executing
fastjsonschema
folium
fonttools
fqdn
google-api-core
google-auth
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-core
google-cloud-resource-manager
google-cloud-storage
google-crc32c
google-genai
google-resumable-media
googleapis-common-protos
grpc-google-iam-v1
grpcio
grpcio-status
h11
httpcore
httpx
idna
importlib-metadata
ipykernel
ipython
ipywidgets
isoduration
jedi
Jinja2
jiter
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kaggle
kiwisolver
latex2mathml
lxml
mammoth
MarkupSafe
matplotlib
matplotlib-inline
mistune
narwhals
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numpy
opencensus
opencensus-context
opencensus-ext-logging
openpyxl
outcome
overrides
packaging
pandas
pandocfilters
parso
pefile
pillow
platformdirs
plotly
prometheus_client
prompt_toolkit
proto-plus
protobuf
psutil
psycopg2
pure-eval
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pygame
Pygments
pyinstaller
pyinstaller-hooks-contrib
pylatexenc
Pympler
pyparsing
PyPDF2
PyQt5
PyQt5-Qt5
PyQt5_sip
PyQtWebEngine
PyQtWebEngine-Qt5
PySocks
python-dateutil
python-docx
python-dotenv
python-json-logger
python-slugify
pytz
pywin32
pywin32-ctypes
pywinpty
PyYAML
pyzmq
qtconsole
QtPy
redis
referencing
requests
rfc3339-validator
rfc3986-validator
rpds-py
rsa
scikit-learn
scipy
seaborn
selenium
Send2Trash
shapely
six
sniffio
sortedcontainers
soupsieve
sqlparse
stack-data
tabula-py
tabulate
tenacity
terminado
text-unidecode
threadpoolctl
tinycss2
tornado
tqdm
traitlets
trio
trio-websocket
types-python-dateutil
typing-inspection
typing_extensions
tzdata
uri-template
urllib3
uv
vertexai
wcwidth
webcolors
webdriver-manager
webencodings
websocket-client
websockets
widgetsnbextension
wrapt
wsproto
xlsxwriter
xyzservices
zipp